Feat/main

See merge request s11-webmobile4-sub2/S11P12D211!91

+++ android/Signal/app/src/main/java/com/ongo/signal/data/model/main/UpdateBoardDTO.kt
@@ -3,5 +3,5 @@ package com.ongo.signal.data.model.main
 data class UpdateBoardDTO(
     val title: String,
     val content: String,
-    val type: Int
+    val type: Long
 )
